feat(api): add OpenAPI documentation to entities
- Add Swagger/OpenAPI annotations to domain entities
  - Product and related DTOs
  - User and related DTOs
  - Request/Response models

- Add detailed field documentation
  - Add descriptive comments for each field
  - Include meaningful example values
  - Document nullable fields in update DTOs

- Improve type consistency
  - Use rust_decimal::Decimal for monetary values
  - Add proper schema examples for all fields
  - Ensure consistent type usage across DTOs

Technical details:
- Add ToSchema derive macro for OpenAPI generation
- Add field-level documentation with /// comments
- Add schema examples with #[schema(example = '..')] attributes

+++ src/domain/entities/product.rs
@@ -3,30 +3,61 @@ use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 use uuid::Uuid;
+use utoipa::ToSchema;
 
-#[derive(Debug, Serialize, Deserialize, FromRow)]
+#[derive(Debug, Serialize, Deserialize, FromRow, ToSchema)]
 pub struct Product {
+    /// The unique identifier for the product
+    #[schema(example = "123e4567-e89b-12d3-a456-426614174000")]
     pub id: Uuid,
+    /// The name of the product
+    #[schema(example = "iPhone 14 Pro")]
     pub name: String,
+    /// Detailed description of the product
+    #[schema(example = "Latest iPhone model with dynamic island")]
     pub description: String,
+    /// The price of the product
+    #[schema(example = "999.99")]
     pub price: Decimal,
+    /// Current stock quantity
+    #[schema(example = "100")]
     pub stock: i32,
+    /// When the product was created
+    #[schema(example = "2024-02-16T00:00:00Z")]
     pub created_at: DateTime<Utc>,
+    /// Last update timestamp
+    #[schema(example = "2024-02-16T00:00:00Z")]
     pub updated_at: DateTime<Utc>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct CreateProductDto {
+    /// The name of the product
+    #[schema(example = "iPhone 14 Pro")]
     pub name: String,
+    /// Detailed description of the product
+    #[schema(example = "Latest iPhone model with dynamic island")]
     pub description: String,
+    /// The price of the product
+    #[schema(example = "999.99")]
     pub price: Decimal,
+    /// Initial stock quantity
+    #[schema(example = "100")]
     pub stock: i32,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct UpdateProductDto {
+    /// Updated name of the product
+    #[schema(example = "iPhone 14 Pro Max")]
     pub name: Option<String>,
+    /// Updated description of the product
+    #[schema(example = "Updated description for iPhone")]
     pub description: Option<String>,
-    pub price: Option<f64>,
+    /// Updated price of the product
+    #[schema(example = "1099.99")]
+    pub price: Option<Decimal>,
+    /// Updated stock quantity
+    #[schema(example = "50")]
     pub stock: Option<i32>,
 }
